# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/books/{book}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a specific book by ID",
                "operationId": "c977385c32095dec31075ca1f77c3cd7",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "The ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A single book",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update a book",
                "operationId": "66285ba337428238f9ea30fdca856ef6",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "The ID of the book to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "isbn": {
                                        "type": "string"
                                    },
                                    "published_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "author_id": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Available",
                                            "Borrowed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a book",
                "operationId": "5bf2d40ecbd2142dabcca0ec2b075369",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "The ID of the book to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book deleted"
                    }
                }
            }
        },
        "/api/books": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "operationId": "72f82e1cd6f08112323f87053277ca99",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "isbn",
                                    "published_date",
                                    "author_id",
                                    "status"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "isbn": {
                                        "type": "string"
                                    },
                                    "published_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "author_id": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Available",
                                            "Borrowed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books/{book}/borrow": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Borrow a book",
                "operationId": "6c775612983471ed15b583d8f8c2e451",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "The ID of the book to borrow",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Book borrowed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BorrowRecord"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books/{book}/return": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Return a borrowed book",
                "operationId": "7ba17ec293352a14e5e820ec68c899d6",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "The ID of the book to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BorrowRecord"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Book": {
                "properties": {
                    "id": {
                        "description": "Book ID",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Book title",
                        "type": "string"
                    },
                    "isbn": {
                        "description": "ISBN",
                        "type": "string"
                    },
                    "author_id": {
                        "description": "Author ID",
                        "type": "integer"
                    },
                    "published_date": {
                        "description": "Published date",
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "description": "Book status",
                        "type": "string",
                        "enum": [
                            "Available",
                            "Borrowed"
                        ]
                    }
                },
                "type": "object"
            },
            "BorrowRecord": {
                "properties": {
                    "id": {
                        "description": "Record ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "description": "User ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "book_id": {
                        "description": "Book ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "borrowed_at": {
                        "description": "Borrowed date and time",
                        "type": "string",
                        "format": "date-time"
                    },
                    "due_at": {
                        "description": "Due date and time",
                        "type": "string",
                        "format": "date-time"
                    },
                    "returned_at": {
                        "description": "Return date and time",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}